VitElectron!
```sh
npx create-vite, vitelectron, Others, create-electron-vite, React
```

https://github.com/electron-vite/vite-plugin-electron

Wow, didn't realize that the Vite tooling CLI already has a setup plugin for use with Electron! Extra awesome!

Seeing how this gets setup really helps me click on how I can use TypeScript, with Electron, and have a framework as well. The TS/framework part of Vite wasn't too complex for me, but hooking that together with TS for the Electron main process and HMR for that as well. This allows Vite to manage the running of `electron .` in the CLI, which usually is the main entry point for the dev server.

Having this as a reference definitely helps a lot! I already trimmed out a few other things, mostly the ESLint setup, since I don't use that just yet. I want to get used to it, but I am going to do that on it's own, adding it to an existing project.

I'm curious how Electron ESM can work with this as well, I want to try that after having this already working nicely.