Browser-Like Context Menus
Sindresorhus is such a great dev! Thanks for making this, I always wanted it to be this easy.

https://github.com/sindresorhus/electron-context-menu

The use of `dispose()` here makes me think that a class could be a really interesting way to setup an Electron app, and to use the new TS-support Dispose API that auto-cleans up your resources after you use them. I like that you don't have to have a class for Electron, it is really nice to just build things with plain event handlers and functional calls.