ESM Electron Main Process
Yay! Looks like it's working now.

I rediscovered that I misunderstood that the preload script *must* be CJS, and the other scripts can be either ESM or CJS now. Keeping the preload script as CJS makes things still work, and the rest of the codebase can be ESM.

I want to properly name the preload script with `.cts`, but Vite isn't understanding the file extension correctly, so it's trying to read/inline it as a video.

https://github.com/electron-vite/vite-plugin-electron/issues/186
https://www.electronjs.org/docs/latest/tutorial/esm