Package-Adjustments
For the past few years, I've been using `;` instead of `&&` for my CLI scripts, because I wanted to ensure they would simply work across both Windows and POSIX systems. As long as I knew of, `&&` isn't available in PowerShell, so I had to find a fallback for while programming on Windows. Now I'm mostly on Unix systems, either Mac or Linux, so it doesn't matter as much anymore. That's not why I'm changing though, I still want anyone to be able to build my projects no matter where they are developing things. I'm only going to keep this as/migrate to `&&` because it looks like PowerShell actually received support for it now!

https://www.google.com/search?q=powershell+double+and
https://stackoverflow.com/questions/563600/can-i-get-or-and-to-work-in-powershell
https://superuser.com/questions/464109/replacements-for-ampersand-and-double-ampersand-in-powershell

Sounds like saving exact Electron versions is recommended just in case of any possible issues that may arise. The official docs don't recommend this though, so I'm kind of questioning/skeptical whether that's as much of an issue as it sounds like.

https://zeke.github.io/electron.atom.io/docs/tutorial/electron-versioning/