Link Forwarding!
Figured out how to forward links out of the app, and into the default browser! I like this much more, in that clicking on website links will properly open not in the same app, but in a regular browser.

https://stackoverflow.com/questions/31749625/make-a-link-from-electron-open-in-browser
https://stackoverflow.com/questions/32402327/how-can-i-force-external-links-from-browser-window-to-open-in-a-default-browser
https://www.electronjs.org/docs/latest/api/window-open#native-window-example
https://www.electronjs.org/docs/latest/breaking-changes#removed-webcontents-new-window-event
https://www.electronjs.org/docs/latest/api/web-contents#contentssetwindowopenhandlerhandler

Don't remember if I committed this yet a few days ago, but I've been listening to DTP Addicted, and I really love the sound of it! Now I'm even more excited for the upcoming PowerNerd album, as it sounds like it will be something along those lines maybe.

https://www.youtube.com/playlist?list=OLAK5uy_nJC-R4_UDldj-DstbzUobYKwR1z9phzro