Main Entry HTML Issues
How do you hook up Vite's build output to that of what Electron-Builder packages as part of the Asar file? It looks like it's instead trying to build and use the source files, which don't run correctly in a regular JS context. It looks like the path has to be absolute as well, making it relative kind of screws itself up. I thought I had figured that part out, but I'm not completely sure now actually.

https://stackoverflow.com/questions/41130993/electron-not-allowed-to-load-local-resource
https://www.reddit.com/r/electronjs/comments/dnwenb/not_allowed_to_load_local_resource_error_when/
https://www.electron.build/configuration/configuration.html

Maybe I can look into just using Electron Forge instead, it might be more up to date and ready to handle build files like I'm trying to use.

I also want to remove the bundling part of the main node script, is there a way to do that? I guess for the frontend it's okay, since it doesn't specifically have access to Node modules.