Build Resolution Debug!
YAAAY!!!! Building the app wasn't properly loading the `index.html` root of the app for some reason, realized it was because I goofed up and removed the Electron-Builder config (I thought it was just a demo, you definitely need it for this setup at least since the configuration is a little different), and I changed the path resolution to the HTML file incorrectly at some point, that was part of the issue as well. Now the app's build folder is saved properly, and the configuration is now present again. YYAAAAYY!!!

Check out the links in the previous commit for reference to what I was looking into.