Vanilla Build Execise
Helping figure out how to have an Electron project with just a bundler, and not a dev environment. I originally was of that mindset for the longest time, while getting into web development. I've been trying to broaden my scope to where it's alright to use bigger tools for things, I think it's probably because now I know how they work though, so it doesn't feel as uncomfortable to use.

To help discern what comes from what with the modern tooling nowadays though, I thought it would be good to try help exercise this concept again, I can learn from it as well. I think modern tools are much better than they used to be though, so it's less of something to worry about, at least partially. I still don't like all-or-nothing tools either. If they provide an inlet for the configuration that I need to customize though (Vite does do this :) ), then I think it's alright to use it as an abstraction.

https://www.reddit.com/r/electronjs/comments/1c3zwps/electron_react/

Tried to stick with Solid for this demo, but I'm not completely sure how to configure that at the esbuild level to import from the correct module.

Another thing to note with this is that it fully works with plain JS too, I only used TypeScript because that was already part of this stack of mine. I think I can look into making a pure JSDoc typings commit next too, to make it more vanilla-esque too.

https://www.freecodecamp.org/news/bundle-a-basic-react-application-using-esbuild/
https://esbuild.github.io/getting-started/#build-scripts
https://esbuild.github.io/content-types/#using-jsx-without-react
https://github.com/evanw/esbuild/issues/718
https://github.com/amoutonbrady/esbuild-plugin-solid